+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project are documented in this file.
 
 - Fixed parent `FocusScope` objects stealing the focus from inner `FocusScope`s when clicked.
 - Added `TextInputInterface.text-input-focused` to detect when a virtual keyboard should open
+- Added `always-on-top` property of a `Window` to show the window above others
 
 ### C++